Refactor and fix lint
